Moved Editor folder into components
